удаление лишнего
