+++ fullstack-frontend-react/src/pages/MainPage.tsx
@@ -1,27 +1,12 @@
-import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-
-import { getCommon } from "../api/controllers/controller";
 
 import {
   Box,
-  Button,
-  Typography
+  Button
 } from "@mui/material";
 
 const MainPage = () => {
   const navigate = useNavigate();
-
-  const [data, setData] = React.useState();
-
-  useEffect(() => {
-    getCommon()
-      .then((response) => {
-        console.log(response);
-        setData(response.data.message);
-      })
-      .catch((e) => console.log(e));
-  }, []);
 
   return <Box
     sx={{
@@ -29,21 +14,10 @@ const MainPage = () => {
       m: '0 auto'
     }}
   >
-    <Typography
-      component={'h2'}
-      sx={{
-        fontSize: '22px',
-        my: 2
-      }}
-    >
-      {
-        data ? data : 'Данных с сервера нет'
-      }
-    </Typography>
     <Button
-    variant='outlined'
-    color='error'
-    onClick={() => navigate('/change-data-backend')}
+      variant='outlined'
+      color='error'
+      onClick={() => navigate('/change-data-backend')}
     >
       Server
     </Button>
